requirements.txt
Django==1.9.2
django-simple-history==1.8.0
django-tastypie==0.13.3
gunicorn==19.4.5
ipaddress==1.0.16
mysqlclient==1.3.7
python-dateutil==2.4.2
python-mimeparse==1.5.1
PyYAML==3.11
requests==2.9.1
simplejson==3.8.2
six==1.10.0
